requirements-sample.txt
aiohttp==3.9.4
aiosignal==1.3.1
anyio==3.6.1
arrow==1.3.0
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
canvasapi==3.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.5
cryptography==42.0.8
dataclasses-json==0.6.4
distro==1.9.0
fastapi==0.70.1
fastapi-health==0.4.0
fastapi-redis-session==0.2.0
frozenlist==1.4.1
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.0
googleapis-common-protos==1.63.0
greenlet==1.1.2
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.13.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.3
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-google-genai==1.0.2
langchain-text-splitters==0.0.1
langsmith==0.1.46
loguru==0.7.2
MarkupSafe==2.1.1
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.17.0
orjson==3.10.0
packaging==23.2
pandas==2.2.2
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==1.10.15
PyJWT==2.8.0
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2024.1
PyYAML==6.0.1
redis==3.5.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==2.0.29
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.16
starlette==0.16.0
tenacity==8.2.3
toolz==0.12.0
tqdm==4.66.2
types-python-dateutil==2.9.0.20240821
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.18.2
yarl==1.9.4
